+++ attributes/default.rb
@@ -2,10 +2,8 @@
 # Cookbook:: composer
 # Attributes:: default
 #
-# Copyright:: 2016-2021, David Joos
+# Copyright:: 2016-2022, David Joos
 #
-
-include_attribute 'php'
 
 if platform?('windows')
   default['composer']['url'] = 'https://getcomposer.org/Composer-Setup.exe'
@@ -26,5 +24,4 @@ end
 
 default['composer']['global_configs'] = {}
 default['composer']['home_dir'] = nil
-default['composer']['php_recipe'] = 'php::default'
 default['composer']['self_update_channel'] = nil